refactor: duration double 타입으로 통일

+++ src/main/java/com/swmStrong/demo/domain/leaderboard/repository/LeaderboardRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.redis.core.ZSetOperations;
 import java.util.Set;
 
 public interface LeaderboardRepository {
-    void increaseScoreByUserId(String key, String userId, long seconds);
+    void increaseScoreByUserId(String key, String userId, double seconds);
     Set<ZSetOperations.TypedTuple<String>> getTopUsers(String key, int topN);
     Long getRankByUserId(String key, String userId);
     Double getScoreByUserId(String key, String userId);
