+++ src/main/java/com/swmStrong/demo/domain/usageLog/dto/CategoryUsageDto.java
@@ -2,7 +2,7 @@ package com.swmStrong.demo.domain.usageLog.dto;
 
 public record CategoryUsageDto(
         String category,
-        long duration,
+        double duration,
         String color
 ) {
 }
