+++ src/main/java/com/swmStrong/demo/message/dto/LeaderBoardUsageMessage.java
@@ -8,7 +8,7 @@ import java.time.LocalDateTime;
 public record LeaderBoardUsageMessage(
         String userId,
         String category,
-        long duration,
+        double duration,
         LocalDateTime timestamp
 ) {
 }
